# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/job_offers:
    get:
      tags:
        - JobOffers
      summary: List JobOffers
      description: Get a paginated list of JobOffers with optional filtering and sorting
      operationId: listJobOffers
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobOfferList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - JobOffers
      summary: Create JobOffer
      description: Create a new JobOffer
      operationId: createJobOffer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateJobOfferRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobOffer'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/job_requisitions:
    get:
      tags:
        - JobRequisitions
      summary: List JobRequisitions
      description: Get a paginated list of JobRequisitions with optional filtering and sorting
      operationId: listJobRequisitions
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobRequisitionList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - JobRequisitions
      summary: Create JobRequisition
      description: Create a new JobRequisition
      operationId: createJobRequisition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateJobRequisitionRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobRequisition'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    JobOffer:
      type: object
      required:
        - id
        - application_id
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        application_id:
          type: string
          format: uuid
        proposed_salary:
          nullable: true
          type: string
        employment_type:
          nullable: true
          type: string
        letter_template_id:
          nullable: true
          type: string
          format: uuid
        status:
          type: string
        offered_at:
          nullable: true
          type: string
          format: date-time
        accepted_at:
          nullable: true
          type: string
          format: date-time
        metadata:
          type: object
...
    CreateJobOfferRequest:
      type: object
      required:
        - application_id
        - status
        - metadata
      properties:
        application_id:
          type: string
          format: uuid
        proposed_salary:
          nullable: true
          type: string
        employment_type:
          nullable: true
          type: string
        letter_template_id:
          nullable: true
          type: string
          format: uuid
        status:
          type: string
        offered_at:
          nullable: true
          type: string
          format: date-time
        accepted_at:
          nullable: true
          type: string
          format: date-time
        metadata:
          type: object
...
    JobOfferList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/JobOffer'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    JobRequisition:
      type: object
      required:
        - id
        - title
        - headcount
        - filled_headcount
        - status
        - opened_by
        - metadata
      properties:
        id:
          type: string
          format: uuid
        department_id:
          nullable: true
          type: string
          format: uuid
        position_id:
          nullable: true
          type: string
          format: uuid
        title:
          type: string
        headcount:
          type: integer
          format: int32
        filled_headcount:
          type: integer
          format: int32
        employment_type:
          nullable: true
          type: string
        status:
          type: string
        opened_by:
          type: string
          format: uuid
        budget:
          nullable: true
          type: string
        deadline:
          nullable: true
          type: string
          format: date
        metadata:
          type: object
...
    CreateJobRequisitionRequest:
      type: object
      required:
        - title
        - headcount
        - filled_headcount
        - status
        - opened_by
        - metadata
      properties:
        department_id:
          nullable: true
          type: string
          format: uuid
        position_id:
          nullable: true
          type: string
          format: uuid
        title:
          type: string
        headcount:
          type: integer
          format: int32
        filled_headcount:
          type: integer
          format: int32
        employment_type:
          nullable: true
          type: string
        status:
          type: string
        opened_by:
          type: string
          format: uuid
        budget:
          nullable: true
          type: string
        deadline:
          nullable: true
          type: string
          format: date
        metadata:
          type: object
...
    JobRequisitionList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/JobRequisition'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []